fix the Distance impl for Point; needed to use an exponent type compatible with both i32 and f32

+++ crates/building_blocks_core/src/point2.rs
@@ -102,7 +102,7 @@ where
 
 impl<T> Distance for Point2<T>
 where
-    T: Copy + Signed + Add<Output = T> + Pow<i32, Output = T>,
+    T: Copy + Signed + Add<Output = T> + Pow<u16, Output = T>,
     Point2<T>: Point<Scalar = T>,
 {
     fn l1_distance(&self, other: &Self) -> Self::Scalar {
